requirements.txt
aiohttp==2.3.2
async-timeout==2.0.0
chardet==3.0.4
docopt==0.6.2
flake8==3.5.0
Logbook==1.1.0
mccabe==0.6.1
msgpack-python==0.4.8
multidict==3.3.2
pkg-resources==0.0.0
pycodestyle==2.3.1
pyflakes==1.6.0
python-dateutil==2.6.1
six==1.11.0
sortedcontainers==1.5.7
ujson==1.35
uvloop==0.8.1
websockets==4.0.1
yarl==0.14.2
